pyproject.toml
[build-system]
requires      = ["setuptools>=61.0.0", "wheel", "pandas"]
build-backend = "setuptools.build_meta"

[project]
name = "digitalmodel"
version = "0.0.7"
description = "engineering asset digital model(s) for life cycle analysis"
readme = "README.md"
authors = [{ name = "Vamsee Achanta", email = "vamsee.achanta@aceengineer.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords=['finite element', 'asset representation', 'structural', 'physical', 'engineering', 'animation', 'lifecycle', 'lifecycle analysis', 'single source of truth', 'digital model', 'digital twin']
dependencies = ["pyyaml", "plotly", "webcolors", "flask_wtf", "flask_httpauth", "flask_restful", "urllib3", 'tomli', 'dash','bumpver','matplotlib','scipy','pandas','xlrd','openpyxl','imgkit','xmltodict','OrcFxAPI; python_version < "3.10"','docx','deepdiff','pytest','sympy','kaleido','assetutilities']
requires-python = ">=3.7"

    [project.optional-dependencies]
    build = ["build", "twine"]
    dev   = ["black", "bumpver", "isort", "pytest"]

    [project.scripts]
    digital_model = "digitalmodel:__main__"

    [project.urls]
    repository    = "https://github.com/vamseeachanta/digitalmodel"
    documentation = "https://github.com/vamseeachanta/digitalmodel/blob/master/README.md"

[tool.setuptools]
include-package-data = true
[tool.setuptools.packages.find]
where = ["src"]
[tool.setuptools.package-data]
"*" = ["*.yml", "*.csv", "*.xlsx", "*.html", "*.txt"]

[tool.bumpver]
current_version = "0.0.7"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message  = "bump version {old_version} -> {new_version}"
commit          = true
tag             = true
push            = false

    [tool.bumpver.file_patterns]
    "pyproject.toml" = [
        'current_version = "{version}"',
        'version = "{version}"',
    ]
    "src/digitalmodel/__init__.py" = ["{version}"]
    "src/digitalmodel/__main__.py" = ["- digitalmodel v{version}"]

[tool.isort]
profile                   = "black"
import_heading_stdlib     = "Standard library imports"
import_heading_thirdparty = "Third party imports"
import_heading_firstparty = "Reader imports"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
strict = true

    [[tool.mypy.overrides]]
    module                 = "feedparser"
    ignore_missing_imports = true


[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''